+++ packages/plugin-ext/src/common/plugin-api-rpc.ts
@@ -1789,6 +1789,7 @@ export type CommentThreadChanges = Partial<{
     contextValue: string,
     comments: Comment[],
     collapseState: CommentThreadCollapsibleState;
+    canReply: boolean;
 }>;
 
 export interface CommentsMain {