+++ packages/plugin-ext/src/main/browser/comments/comment-thread-widget.tsx
@@ -89,6 +89,12 @@ export class CommentThreadWidget extends BaseWidget {
                 commentForm.update();
             }
         }));
+        this.toDispose.push(this._commentThread.onDidChangeCanReply(_canReply => {
+            const commentForm = this.commentFormRef.current;
+            if (commentForm) {
+                commentForm.update();
+            }
+        }));
         this.contextMenu = this.menus.getMenu(COMMENT_THREAD_CONTEXT);
         this.contextMenu.children.map(node => node instanceof ActionMenuNode && node.action.when).forEach(exp => {
             if (typeof exp === 'string') {
@@ -381,7 +387,7 @@ export class CommentForm<P extends CommentForm.Props = CommentForm.Props> extend
     override render(): React.ReactNode {
         const { commands, commentThread, contextKeyService } = this.props;
         const hasExistingComments = commentThread.comments && commentThread.comments.length > 0;
-        return <div className={'comment-form' + (this.state.expanded || commentThread.comments && commentThread.comments.length === 0 ? ' expand' : '')}>
+        return commentThread.canReply ? <div className={'comment-form' + (this.state.expanded || commentThread.comments && commentThread.comments.length === 0 ? ' expand' : '')}>
             <div className={'theia-comments-input-message-container'}>
                 <textarea className={'theia-comments-input-message theia-input'}
                     spellCheck={false}
@@ -409,7 +415,7 @@ export class CommentForm<P extends CommentForm.Props = CommentForm.Props> extend
                 clearInput={this.clearInput}
             />
             <button className={'review-thread-reply-button'} title={'Reply...'} onClick={this.expand}>Reply...</button>
-        </div>;
+        </div> : null;
     }
 }